TST: de-duplicate PeriodIndex constructor tests (#56435)
* rename test_period_asfreq->test_period

* TST: de-duplicate PeriodIndex constructor tests
