+++ pandas/tests/indexes/period/test_period.py
@@ -1,8 +1,6 @@
 import numpy as np
 import pytest
 
-from pandas._libs.tslibs.period import IncompatibleFrequency
-
 from pandas import (
     Index,
     NaT,
@@ -17,11 +15,6 @@ import pandas._testing as tm
 
 
 class TestPeriodIndex:
-    def test_make_time_series(self):
-        index = period_range(freq="Y", start="1/1/2001", end="12/1/2009")
-        series = Series(1, index=index)
-        assert isinstance(series, Series)
-
     def test_view_asi8(self):
         idx = PeriodIndex([], freq="M")
 
@@ -65,74 +58,6 @@ class TestPeriodIndex:
         tm.assert_numpy_array_equal(idx.to_numpy(), exp)
         exp = np.array([14975, -9223372036854775808], dtype=np.int64)
         tm.assert_numpy_array_equal(idx.asi8, exp)
-
-    def test_period_index_length(self):
-        pi = period_range(freq="Y", start="1/1/2001", end="12/1/2009")
-        assert len(pi) == 9
-
-        pi = period_range(freq="Q", start="1/1/2001", end="12/1/2009")
-        assert len(pi) == 4 * 9
-
-        pi = period_range(freq="M", start="1/1/2001", end="12/1/2009")
-        assert len(pi) == 12 * 9
-
-        msg = "Period with BDay freq is deprecated"
-        with tm.assert_produces_warning(FutureWarning, match=msg):
-            start = Period("02-Apr-2005", "B")
-            i1 = period_range(start=start, periods=20)
-        assert len(i1) == 20
-        assert i1.freq == start.freq
-        assert i1[0] == start
-
-        end_intv = Period("2006-12-31", "W")
-        i1 = period_range(end=end_intv, periods=10)
-        assert len(i1) == 10
-        assert i1.freq == end_intv.freq
-        assert i1[-1] == end_intv
-
-        end_intv = Period("2006-12-31", "1w")
-        i2 = period_range(end=end_intv, periods=10)
-        assert len(i1) == len(i2)
-        assert (i1 == i2).all()
-        assert i1.freq == i2.freq
-
-        msg = "start and end must have same freq"
-        msg2 = "Period with BDay freq is deprecated"
-        with pytest.raises(ValueError, match=msg):
-            with tm.assert_produces_warning(FutureWarning, match=msg2):
-                period_range(start=start, end=end_intv)
-
-        with tm.assert_produces_warning(FutureWarning, match=msg2):
-            end_intv = Period("2005-05-01", "B")
-        with tm.assert_produces_warning(FutureWarning, match=msg2):
-            i1 = period_range(start=start, end=end_intv)
-
-        msg = (
-            "Of the three parameters: start, end, and periods, exactly two "
-            "must be specified"
-        )
-        with pytest.raises(ValueError, match=msg):
-            period_range(start=start)
-
-        # infer freq from first element
-        with tm.assert_produces_warning(FutureWarning, match=msg2):
-            i2 = PeriodIndex([end_intv, Period("2005-05-05", "B")])
-        assert len(i2) == 2
-        assert i2[0] == end_intv
-
-        with tm.assert_produces_warning(FutureWarning, match=msg2):
-            i2 = PeriodIndex(np.array([end_intv, Period("2005-05-05", "B")]))
-        assert len(i2) == 2
-        assert i2[0] == end_intv
-
-        # Mixed freq should fail
-        vals = [end_intv, Period("2006-12-31", "w")]
-        msg = r"Input has different freq=W-SUN from PeriodIndex\(freq=B\)"
-        with pytest.raises(IncompatibleFrequency, match=msg):
-            PeriodIndex(vals)
-        vals = np.array(vals)
-        with pytest.raises(ValueError, match=msg):
-            PeriodIndex(vals)
 
     @pytest.mark.parametrize(
         "field",
@@ -210,23 +135,6 @@ class TestPeriodIndex:
         tm.assert_index_equal(idx.unique(), expected)
         assert idx.nunique() == 3
 
-    def test_negative_ordinals(self):
-        Period(ordinal=-1000, freq="Y")
-        Period(ordinal=0, freq="Y")
-
-        msg = "The 'ordinal' keyword in PeriodIndex is deprecated"
-        with tm.assert_produces_warning(FutureWarning, match=msg):
-            idx1 = PeriodIndex(ordinal=[-1, 0, 1], freq="Y")
-        with tm.assert_produces_warning(FutureWarning, match=msg):
-            idx2 = PeriodIndex(ordinal=np.array([-1, 0, 1]), freq="Y")
-        tm.assert_index_equal(idx1, idx2)
-
-        alt1 = PeriodIndex.from_ordinals([-1, 0, 1], freq="Y")
-        tm.assert_index_equal(alt1, idx1)
-
-        alt2 = PeriodIndex.from_ordinals(np.array([-1, 0, 1]), freq="Y")
-        tm.assert_index_equal(alt2, idx2)
-
     def test_pindex_fieldaccessor_nat(self):
         idx = PeriodIndex(
             ["2011-01", "2011-02", "NaT", "2012-03", "2012-04"], freq="D", name="name"
@@ -280,50 +188,6 @@ class TestPeriodIndex:
         result = index.map(lambda x: x.ordinal)
         exp = Index([x.ordinal for x in index])
         tm.assert_index_equal(result, exp)
-
-    @pytest.mark.parametrize(
-        "freq,freq_depr",
-        [
-            ("2M", "2ME"),
-            ("2Q-MAR", "2QE-MAR"),
-            ("2Y-FEB", "2YE-FEB"),
-        ],
-    )
-    def test_period_index_frequency_ME_error_message(self, freq, freq_depr):
-        # GH#52064
-        msg = f"for Period, please use '{freq[1:]}' instead of '{freq_depr[1:]}'"
-
-        with pytest.raises(ValueError, match=msg):
-            PeriodIndex(["2020-01-01", "2020-01-02"], freq=freq_depr)
-
-    def test_H_deprecated_from_time_series(self):
-        # GH#52536
-        msg = "'H' is deprecated and will be removed in a future version."
-
-        with tm.assert_produces_warning(FutureWarning, match=msg):
-            index = period_range(freq="2H", start="1/1/2001", end="12/1/2009")
-        series = Series(1, index=index)
-        assert isinstance(series, Series)
-
-    @pytest.mark.parametrize("freq_depr", ["2A", "A-DEC", "200A-AUG"])
-    def test_a_deprecated_from_time_series(self, freq_depr):
-        # GH#52536
-        freq_msg = freq_depr[freq_depr.index("A") :]
-        msg = f"'{freq_msg}' is deprecated and will be removed in a future version, "
-        f"please use 'Y{freq_msg[1:]}' instead."
-
-        with tm.assert_produces_warning(FutureWarning, match=msg):
-            index = period_range(freq=freq_depr, start="1/1/2001", end="12/1/2009")
-        series = Series(1, index=index)
-        assert isinstance(series, Series)
-
-    @pytest.mark.parametrize("freq_depr", ["2SME", "2CBME", "2BYE"])
-    def test_period_index_frequency_invalid_freq(self, freq_depr):
-        # GH#9586
-        msg = f"Invalid frequency: {freq_depr[1:]}"
-
-        with pytest.raises(ValueError, match=msg):
-            period_range("2020-01", "2020-05", freq=freq_depr)
 
 
 def test_maybe_convert_timedelta():
