+++ pandas/tests/indexes/period/test_period_range.py
@@ -11,7 +11,7 @@ from pandas import (
 import pandas._testing as tm
 
 
-class TestPeriodRange:
+class TestPeriodRangeKeywords:
     def test_required_arguments(self):
         msg = (
             "Of the three parameters: start, end, and periods, exactly two "
@@ -20,6 +20,62 @@ class TestPeriodRange:
         with pytest.raises(ValueError, match=msg):
             period_range("2011-1-1", "2012-1-1", "B")
 
+    def test_required_arguments2(self):
+        start = Period("02-Apr-2005", "D")
+        msg = (
+            "Of the three parameters: start, end, and periods, exactly two "
+            "must be specified"
+        )
+        with pytest.raises(ValueError, match=msg):
+            period_range(start=start)
+
+    def test_required_arguments3(self):
+        # not enough params
+        msg = (
+            "Of the three parameters: start, end, and periods, "
+            "exactly two must be specified"
+        )
+        with pytest.raises(ValueError, match=msg):
+            period_range(start="2017Q1")
+
+        with pytest.raises(ValueError, match=msg):
+            period_range(end="2017Q1")
+
+        with pytest.raises(ValueError, match=msg):
+            period_range(periods=5)
+
+        with pytest.raises(ValueError, match=msg):
+            period_range()
+
+    def test_required_arguments_too_many(self):
+        msg = (
+            "Of the three parameters: start, end, and periods, "
+            "exactly two must be specified"
+        )
+        with pytest.raises(ValueError, match=msg):
+            period_range(start="2017Q1", end="2018Q1", periods=8, freq="Q")
+
+    def test_start_end_non_nat(self):
+        # start/end NaT
+        msg = "start and end must not be NaT"
+        with pytest.raises(ValueError, match=msg):
+            period_range(start=NaT, end="2018Q1")
+        with pytest.raises(ValueError, match=msg):
+            period_range(start=NaT, end="2018Q1", freq="Q")
+
+        with pytest.raises(ValueError, match=msg):
+            period_range(start="2017Q1", end=NaT)
+        with pytest.raises(ValueError, match=msg):
+            period_range(start="2017Q1", end=NaT, freq="Q")
+
+    def test_periods_requires_integer(self):
+        # invalid periods param
+        msg = "periods must be a number, got foo"
+        with pytest.raises(TypeError, match=msg):
+            period_range(start="2017Q1", periods="foo")
+
+
+class TestPeriodRange:
     @pytest.mark.parametrize(
         "freq_offset, freq_period",
         [
@@ -95,7 +151,8 @@ class TestPeriodRange:
         tm.assert_index_equal(result, expected)
 
         # downsampling
-        start, end = Period("2017-1", freq="M"), Period("2019-12", freq="M")
+        start = Period("2017-1", freq="M")
+        end = Period("2019-12", freq="M")
         expected = date_range(
             start="2017-01-31", end="2019-12-31", freq="QE", name="foo"
         ).to_period()
@@ -103,7 +160,8 @@ class TestPeriodRange:
         tm.assert_index_equal(result, expected)
 
         # test for issue # 21793
-        start, end = Period("2017Q1", freq="Q"), Period("2018Q1", freq="Q")
+        start = Period("2017Q1", freq="Q")
+        end = Period("2018Q1", freq="Q")
         idx = period_range(start=start, end=end, freq="Q", name="foo")
         result = idx == idx.values
         expected = np.array([True, True, True, True, True])
@@ -121,42 +179,43 @@ class TestPeriodRange:
         result = period_range(start=end, end=start, freq="W", name="foo")
         tm.assert_index_equal(result, expected)
 
-    def test_errors(self):
-        # not enough params
+    def test_mismatched_start_end_freq_raises(self):
+        depr_msg = "Period with BDay freq is deprecated"
+        end_w = Period("2006-12-31", "1w")
+
+        with tm.assert_produces_warning(FutureWarning, match=depr_msg):
+            start_b = Period("02-Apr-2005", "B")
+            end_b = Period("2005-05-01", "B")
+
+        msg = "start and end must have same freq"
+        with pytest.raises(ValueError, match=msg):
+            with tm.assert_produces_warning(FutureWarning, match=depr_msg):
+                period_range(start=start_b, end=end_w)
+
+        # without mismatch we are OK
+        with tm.assert_produces_warning(FutureWarning, match=depr_msg):
+            period_range(start=start_b, end=end_b)
+
+
+class TestPeriodRangeDisallowedFreqs:
+    def test_constructor_U(self):
+        # U was used as undefined period
+        with pytest.raises(ValueError, match="Invalid frequency: X"):
+            period_range("2007-1-1", periods=500, freq="X")
+
+    def test_H_deprecated_from_time_series(self):
+        # GH#52536
+        msg = "'H' is deprecated and will be removed in a future version."
+        with tm.assert_produces_warning(FutureWarning, match=msg):
+            period_range(freq="2H", start="1/1/2001", end="12/1/2009")
+
+    @pytest.mark.parametrize("freq_depr", ["2A", "A-DEC", "200A-AUG"])
+    def test_a_deprecated_from_time_series(self, freq_depr):
+        # GH#52536
+        freq_msg = freq_depr[freq_depr.index("A") :]
         msg = (
-            "Of the three parameters: start, end, and periods, "
-            "exactly two must be specified"
+            f"'{freq_msg}' is deprecated and will be removed in a future version, "
+            f"please use 'Y{freq_msg[1:]}' instead."
         )
-        with pytest.raises(ValueError, match=msg):
-            period_range(start="2017Q1")
-
-        with pytest.raises(ValueError, match=msg):
-            period_range(end="2017Q1")
-
-        with pytest.raises(ValueError, match=msg):
-            period_range(periods=5)
-
-        with pytest.raises(ValueError, match=msg):
-            period_range()
-
-        # too many params
-        with pytest.raises(ValueError, match=msg):
-            period_range(start="2017Q1", end="2018Q1", periods=8, freq="Q")
-
-        # start/end NaT
-        msg = "start and end must not be NaT"
-        with pytest.raises(ValueError, match=msg):
-            period_range(start=NaT, end="2018Q1")
-
-        with pytest.raises(ValueError, match=msg):
-            period_range(start="2017Q1", end=NaT)
-
-        # invalid periods param
-        msg = "periods must be a number, got foo"
-        with pytest.raises(TypeError, match=msg):
-            period_range(start="2017Q1", periods="foo")
-
-    def test_period_range_frequency_ME_error_message(self):
-        msg = "for Period, please use 'M' instead of 'ME'"
-        with pytest.raises(ValueError, match=msg):
-            period_range(start="Jan-2000", end="Dec-2000", freq="2ME")
+        with tm.assert_produces_warning(FutureWarning, match=msg):
+            period_range(freq=freq_depr, start="1/1/2001", end="12/1/2009")
