+++ pandas/tests/scalar/period/test_period.py
@@ -7,10 +7,7 @@ from datetime import (
 import numpy as np
 import pytest
 
-from pandas._libs.tslibs import (
-    iNaT,
-    period as libperiod,
-)
+from pandas._libs.tslibs import iNaT
 from pandas._libs.tslibs.ccalendar import (
     DAYS,
     MONTHS,
@@ -31,7 +28,22 @@ import pandas._testing as tm
 bday_msg = "Period with BDay freq is deprecated"
 
 
-class TestPeriodConstruction:
+class TestPeriodDisallowedFreqs:
+    @pytest.mark.parametrize(
+        "freq, freq_msg",
+        [
+            (offsets.BYearBegin(), "BYearBegin"),
+            (offsets.YearBegin(2), "YearBegin"),
+            (offsets.QuarterBegin(startingMonth=12), "QuarterBegin"),
+            (offsets.BusinessMonthEnd(2), "BusinessMonthEnd"),
+        ],
+    )
+    def test_offsets_not_supported(self, freq, freq_msg):
+        # GH#55785
+        msg = f"{freq_msg} is not supported as period frequency"
+        with pytest.raises(TypeError, match=msg):
+            Period(year=2014, freq=freq)
+
     def test_custom_business_day_freq_raises(self):
         # GH#52534
         msg = "CustomBusinessDay is not supported as period frequency"
@@ -40,6 +52,18 @@ class TestPeriodConstruction:
         with pytest.raises(TypeError, match=msg):
             Period("2023-04-10", freq=offsets.CustomBusinessDay())
 
+    def test_invalid_frequency_error_message(self):
+        msg = "WeekOfMonth is not supported as period frequency"
+        with pytest.raises(TypeError, match=msg):
+            Period("2012-01-02", freq="WOM-1MON")
+
+    def test_invalid_frequency_period_error_message(self):
+        msg = "for Period, please use 'M' instead of 'ME'"
+        with pytest.raises(ValueError, match=msg):
+            Period("2012-01-02", freq="ME")
+
+
+class TestPeriodConstruction:
     def test_from_td64nat_raises(self):
         # GH#44507
         td = NaT.to_numpy("m8[ns]")
@@ -114,9 +138,13 @@ class TestPeriodConstruction:
         with pytest.raises(ValueError, match=msg):
             Period("2007-1-1", freq="X")
 
+    def test_tuple_freq_disallowed(self):
         # GH#34703 tuple freq disallowed
         with pytest.raises(TypeError, match="pass as a string instead"):
             Period("1982", freq=("Min", 1))
+
+        with pytest.raises(TypeError, match="pass as a string instead"):
+            Period("2006-12-31", ("w", 1))
 
     def test_construction_from_timestamp_nanos(self):
         # GH#46811 don't drop nanos from Timestamp
@@ -1074,13 +1102,6 @@ class TestPeriodProperties:
         )
 
 
-class TestPeriodField:
-    def test_get_period_field_array_raises_on_out_of_range(self):
-        msg = "Buffer dtype mismatch, expected 'const int64_t' but got 'double'"
-        with pytest.raises(ValueError, match=msg):
-            libperiod.get_period_field_arr(-1, np.empty(1), 0)
-
-
 class TestPeriodComparisons:
     def test_sort_periods(self):
         jan = Period("2000-01", "M")
@@ -1127,15 +1148,3 @@ def test_negone_ordinals():
     repr(period)
     period = Period(ordinal=-1, freq="W")
     repr(period)
-
-
-def test_invalid_frequency_error_message():
-    msg = "WeekOfMonth is not supported as period frequency"
-    with pytest.raises(TypeError, match=msg):
-        Period("2012-01-02", freq="WOM-1MON")
-
-
-def test_invalid_frequency_period_error_message():
-    msg = "for Period, please use 'M' instead of 'ME'"
-    with pytest.raises(ValueError, match=msg):
-        Period("2012-01-02", freq="ME")
